Refactor codebase
Having looked at Google's stylesheet for shell code,
my intentions were to refactor the codebase based on
their suggestions and my own style. This first part was to focus
mainly on the shell code but to also expand the repo
with a makefile.

Scripts have been moved into respective dirs depending on
the language they been written in. Makefile has been added
that will function the same as lower-all-files did.

Some scripts have been deleted if I did not feel as if it would
be worth maintaining them. update-remote-forks is currently
undergoing to be converted into a python script due to growing
complexity.