Refactor genconfigs.py
Several LBYL instances have been replaced with EAFP expressions.

Docstrings have been refactored, w/the extended summary for the
class docstring being moved into as the module level docstring.
A lot of the class variables now contain a prefixed underscore. This
was mainly done to avoid having to write about each class attribute as
part of the class docstring.

Finally, a script name now does not have to be passed in to flatten
the configuration datastructure. It is optional todo so if not wanting
to use the common configuration prefix to denote that a script uses
a common configuration in the root of the configuration datastructure.