Reimplement genconfigs as a shell script
This is mainly so in the event I'd only wish to use a particular
languages's scripts, that I wouldn't be tied down to having to install
python (and its environment) in conjunction with the other language.

For now, the usage of having a duplicate configuration in a particular
script has not been added back. I have not used this feature but if
desired I will bring it back into the shell implemented genconfigs.