Redo target structure for makefile
Main reason is because grouping generic commands that all language
specific targets should run regardless which is chosen would not be
intuitive to use and work off of given the current structure of the
targets.

Instead, I decided restructure the targets to allow a particular make
configuration determine what set of scripts an action is performed on.
This allows me define a generic set of commands more intuitively and
with somewhat better scale.

Inspired by:
https://stackoverflow.com/questions/7656425/makefile-ifeq-logical-or