Have python shims append pyenv paths to PATH
This is to get around the limitation of assuming the calling shell has
all the needed pyenv environment variables set.

With Jenkins jobs (including pipelines), no interactive shell is called
on. Meaning profile and .bashrc are not sourced when running a shell
step in a Jenkins job. My desires are to not focus on altering the
Jenkins instance configuration or jobs themselves to make up for this.
The jobs/instance should assume the build node/project will take care of
these details.