Add notion of a active config for the project
This is to help take guess work out for the project's Jenkins pipeline steps
when needing to place a configuration file prior to even creating the
project's pipeline on Jenkins. This because genconfigs will set a active
configuration for each sequential invocation after the initial time
genconfigs is invoked.

The path to the active configuration can be set now either with a option
to genconfigs or through an environment variable.