Change parameters for disable_github_actions.py
The intent of the script was to disable GitHub forked repos that I did not
want actions running for. Hence I can assume for that I will be the
sole owner of the repos. That being said, the script should technically
work with any GitHub repo that I can particular permissions to (at
least to modify workflows of the repo).

Instead, I'd rather list the names of repos I want to disable actions
for and use the previously defined github_username configuration
that already exists as part of the configuration definition.