Remove exiting w/status code 1 for SystemExits
All my python scripts for retrieve_cmd_args return an exit status code
of 1 in the event ArgumentParser.parse_args finds an invalid option.
However, this also extends to the -h/--help option. Which shouldn't be
the case as -h/--help is a valid option that can be passed in.