Rework update_github_forks.py
The tool.black section has been added to pyproject.toml as this is the
recommended (per black's documentation) way to add custom black
configurations.

The TokenAuth class has been added, as this still seems to be the
recommended way to authenticate to GitHub with a Bearer token using the
requests library. The implementation of the class derived from
https://realpython.com/python-requests/#authentication.

disable_github_actions.py has been combined into update_gh_forks.py as
the only workflows I currently want to disable are those from forked
repositories.

The forked repositories configuration has been moved within
update_gh_forks.py because no other script relies on this configuration
and the configuration won't be different from what's ran in CI/CD. JSON
will no longer be loaded via stdin because I intend to move away from
using genconfigs. The syncing of a branch upstream and disabling of a
workflow are done via the requests library because neither of these
functionalities are exposed for the respective models.

A workflow with the name 'pages-build-deployment' will not be disabled
because this workflow can't be disabled. I believe this to be because
the workflow is tethered to GitHub Pages. Hence, as long as I don't
track that branch on a fork, that workflow shouldn't be triggered. Some
discussion on this can be found here,
https://github.com/quarto-dev/quarto-cli/discussions/5810.