Migrate from Jenkins to GitHub Actions
genconfigs and its related tooling have been removed because I do not
intend on using it any further.

Go binaries are installed within the repository's local bin directory
because this is still a recommended approach by official Golang's
documentation, as mentioned here,
https://go.dev/wiki/Modules#how-can-i-track-tool-dependencies-for-a-module.
The usage of the local bin directory was of my idea, as I would rather
not use the default GOBIN directory. Which I observed could be defined
by the path '${GOPATH}/bin'.