Include scripts within the gps-python image
A bind mount is used for including requirements.txt within the container
because by using a bind mount, requirements.txt will not persist in the
final image. This is considered a best practice from Docker's
documentation.

The WORKDIR instruction argument has been changed so wherever this image
gets used doesn't have to know about specific pathing to execute any of
the scripts.