Include upstream branches to sync new branches
This was done because just iterating on the forked repository's branches
would mean newer branches pushed upstream would never be added to the
forked repository. This also meant that a forked repository could become
empty. Such a situation could occur due to none of the branches in the
forked repository existing upstream, which would lead to all of said
branches being deleted.

The gh-pages branch is filtered out of the upstream branches and not
the forked branches because a forked repository may still have the
branch when initially forked. Meaning, I would want said branch to be
deleted when the script runs on the forked repository for the first
time.

A new branch is created if it cannot be found in the forked repository
because the branch must exist to be merged with its upstream
counterpart. Pointing the newly created branch to the same commit used
by the upstream branch has the same effect as syncing the branches.