Improve logging within update_gh_forks.py
This was done because the previous amount of logging done was inadequate
for troubleshooting when the script gets stuck. The logger's INFO level
equivalent method was the method added primarily throughout to be able
to confirm "...that things are working as expected" (verbiage per
'logging' library documentation).

The LOGLEVEL environment variable was named after a lowercased variable
of the same name used within the 'logging' library documentation.