Update forked GitHub repositories' concurrently
max_workers was set to 3 for both ThreadPoolExecutors because I found
this number of threads to result in an adequate speedup and to not cause
GitHub's API to start throwing 403s due to the frequency of requests. I
did find that setting max_workers to 4 worked fine; however, the
speedup was only an additional 5-10 seconds. This speedup isn't worth
the additional worker and risk involved with GitHub's API possibly
throwing 403s.

Workflows are managed in a separate executor block because I want to
work on them only after the branches' executor block has been run. Newly
created or synced branches can have running workflows that should be
disabled.